#17 added mixins for buttons, containerCards and statistics. Moved thread Components to seperate folder
