+++ src/assets/css/app.scss
@@ -29,6 +29,9 @@ html,
 body {
   height: 100%;
   margin: 0;
+  * {
+    font-family: Arial, Helvetica, sans-serif;
+  }
 }
 
 body {
