+++ src/assets/css/mixins/_statistics.scss
@@ -1,0 +1,17 @@
+@import "../variables";
+
+@mixin statistics() {
+  display: table;
+  text-align: left;
+  align-self: flex-start;
+  .row {
+    display: table-row;
+    span {
+      display: table-cell;
+    }
+
+    .highlighted {
+      color: $highlightedFontColor;
+    }
+  }
+}
